Clientservice: Separate library
This change introduces the clientservice-lib library. This will allow us to
build unit test exectuables.
